# fitnessWebApp-UI/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Andrea Barutta",
    "description": "OpenApi documentation for Spring Security",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Andrea Barutta",
      "url": "https://barutta02.github.io/Barutta_portfolio/",
      "email": "andreabarutta123@gmail.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "Local ENV"
    },
    {
      "url": "https://some-url.com/api/v1",
      "description": "PROD ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/periods": {
      "post": {
        "tags": [
          "Periods"
        ],
        "operationId": "savePeriodo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeriodoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeriodoResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/exercises": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "operationId": "findAllExercise",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseExerciseResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Exercise"
        ],
        "operationId": "saveExercise",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/exercises/import/{exercise-id}": {
      "post": {
        "tags": [
          "Exercise"
        ],
        "operationId": "importExercise",
        "parameters": [
          {
            "name": "exercise-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/exercises/cover/{exercise-id}": {
      "post": {
        "tags": [
          "Exercise"
        ],
        "operationId": "uploadBookCoverPicture",
        "parameters": [
          {
            "name": "exercise-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/periods/{periodo-id}": {
      "get": {
        "tags": [
          "Periods"
        ],
        "operationId": "findPeriodoById",
        "parameters": [
          {
            "name": "periodo-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeriodoResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/periods/creator": {
      "get": {
        "tags": [
          "Periods"
        ],
        "operationId": "findAllPeriodoByCreator",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponsePeriodoResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/exercises/{exercise-id}": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "operationId": "findExerciseById",
        "parameters": [
          {
            "name": "exercise-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Exercise"
        ],
        "operationId": "deleteExercise",
        "parameters": [
          {
            "name": "exercise-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/exercises/exercise-store": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "operationId": "getExercisesFromPublicStore",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseExerciseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/exercises/creator": {
      "get": {
        "tags": [
          "Exercise"
        ],
        "operationId": "findAllExerciseByCreator",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseExerciseResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/periods/{periodo-nome}": {
      "delete": {
        "tags": [
          "Periods"
        ],
        "operationId": "deletePeriodo",
        "parameters": [
          {
            "name": "periodo-nome",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PeriodoResponse": {
        "required": [
          "attivo",
          "data_inizio",
          "durata_in_giorni",
          "id",
          "name",
          "obiettivo"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "obiettivo": {
            "type": "string",
            "enum": [
              "DIMAGRIMENTO",
              "DEFINIZIONE",
              "MASSA",
              "FORZA",
              "RESISTENZA",
              "NON_DEFINITO"
            ]
          },
          "durata_in_giorni": {
            "type": "integer",
            "format": "int32"
          },
          "data_inizio": {
            "type": "string",
            "format": "date"
          },
          "data_fine": {
            "type": "string",
            "format": "date"
          },
          "attivo": {
            "type": "boolean"
          },
          "creator_username": {
            "type": "string"
          }
        }
      },
...
      "PeriodoRequest": {
        "required": [
          "attivo",
          "data_inizio",
          "durata_in_giorni",
          "name",
          "obiettivo"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "obiettivo": {
            "type": "string",
            "enum": [
              "DIMAGRIMENTO",
              "DEFINIZIONE",
              "MASSA",
              "FORZA",
              "RESISTENZA",
              "NON_DEFINITO"
            ]
          },
          "durata_in_giorni": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "data_inizio": {
            "type": "string",
            "format": "date"
          },
          "data_fine": {
            "type": "string",
            "format": "date"
          },
          "attivo": {
            "type": "boolean"
          }
        }
      },
...
      "ExerciseRequest": {
        "required": [
          "description",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "ABS",
                "ARMS",
                "BACK",
                "CALVES",
                "CHEST",
                "LEGS",
                "CORE",
                "SHOULDERS",
                "CARDIO",
                "FULL_BODY",
                "STRETCHING",
                "YOGA",
                "PILATES",
                "CROSSFIT",
                "FUNCTIONAL",
                "POWERLIFTING",
                "OLYMPIC_LIFTING",
                "STRONGMAN",
                "BODYBUILDING",
                "AEROBICS",
                "ZUMBA",
                "DANCE",
                "MARTIAL_ARTS",
                "NON_DEFINITO",
                "OTHER"
              ]
            }
          },
          "shareable": {
            "type": "boolean"
          }
        }
      },
      "ExerciseResponse": {
        "required": [
          "description",
          "id",
          "name",
          "shareable"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "creator_username": {
            "type": "string"
          },
          "category": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "ABS",
                "ARMS",
                "BACK",
                "CALVES",
                "CHEST",
                "LEGS",
                "CORE",
                "SHOULDERS",
                "CARDIO",
                "FULL_BODY",
                "STRETCHING",
                "YOGA",
                "PILATES",
                "CROSSFIT",
                "FUNCTIONAL",
                "POWERLIFTING",
                "OLYMPIC_LIFTING",
                "STRONGMAN",
                "BODYBUILDING",
                "AEROBICS",
                "ZUMBA",
                "DANCE",
                "MARTIAL_ARTS",
                "NON_DEFINITO",
                "OTHER"
              ]
            }
          },
          "cover": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "byte"
            }
          },
          "shareable": {
            "type": "boolean"
          }
        }
      },
...
      "PageResponsePeriodoResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PeriodoResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
...
      "PageResponseExerciseResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}